requirements.txt
bcrypt==3.2.0
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.5
ipython==8.0.1
mysqlclient==2.1.0
PyJWT==2.3.0
PyMySQL==1.0.2
requests==2.27.1
django-storages==1.12.3
boto3==1.21.1
gunicorn==20.1.0
